TRACK-95 - Refactored logic around students in a lesson from students controller into a student_lessons controller
