+++ app/views/lessons/show.html.erb
@@ -14,9 +14,9 @@
   </thead>
   <% @lesson.group.students.each do |student| %>
     <tr class="resource-row student-grade-row">
-        <td><%= link_to student.last_name, grades_lesson_student_path(@lesson.id, student.id), class:'resource-row-link student-grade-row-link' %></td>
-        <td><%= link_to student.first_name, grades_lesson_student_path(@lesson.id, student.id), class:'resource-row-link student-grade-row-link' %></td>
-        <td class="text-xs-center"><%= link_to lesson_student_average_grade(@lesson, student), grades_lesson_student_path(@lesson.id, student.id),
+        <td><%= link_to student.last_name, lesson_student_path(@lesson.id, student.id), class:'resource-row-link student-grade-row-link' %></td>
+        <td><%= link_to student.first_name, lesson_student_path(@lesson.id, student.id), class:'resource-row-link student-grade-row-link' %></td>
+        <td class="text-xs-center"><%= link_to lesson_student_average_grade(@lesson, student), lesson_student_path(@lesson.id, student.id),
                                                class:'resource-row-link student-grade-row-link' %>
         </td>
     </tr>
